Added persistance functions to store/restore model
This completes the implementation of the Model
interface from the github.com/cdipaolo/goml/base
package. Still need to add functions to turn a
CSV file into a dataset that can be used by the
library, but that should be pretty easy.

Testing is decently rigorous, currently, for the
completed LeastSquares (ordinary LeastSquares, not
completely general) struct/model. All tests follow
an instatiation pattern of creating data that matches
some linear function and then trains it, calling it
at different points and expecting the correct function
value to return. The model has been tested in 1,2,and
3 feature dimensions (not including the constant term,)
and that should be good enough to test for accurate
convergance on all higher orders, though a benchmark
might be needed to test very large datasets and see
how efficient it is with those. The model currently
uses Batch Gradient Descent, though, so it might not
be the best model to use for a huge dataset.