added stochastic gradient descent for optimization
I added optionality within the optimization method
used to optimize the prediction function within both
the Logistic and LeastSquares models. Now, when using
NewLeastSquares or NewLogistic, you set the method
within the first parameter.

For example to use Stochastic Gradient Ascent on an
ordinary LeastSquares model:

    model := NewLeastSquares(base.StochasticGA, 1e-4, 1e3, 300,
        increasingX, increasingY)

To use Batch Gradient Descent on a Logistic model:

    model := NewLogistic(base.StochasticGA, 1e-4, 1e3, 300,
        twoDX, twoDY)

and that's it. Methods are defined within the base
package as a casted string type, defined as follows:

// OptimizationMethod defines a type enum which
// (using constants declared below) lets a user
// pass in a optimization method to use when
// creating a new model
type OptimizationMethod string

const (
    BatchGA OptimizationMethod = "Batch Gradient Ascent"
    StochasticGA = "Stochastic Gradient Descent"
)

The reason they are strings and not an iota defined
enum is because the method casted is used when describing
the model before training in the nice, clean summary,
so you know which method you're using.

I might add in Newton's Method later, but currently
only Stochastic and Batch Gradient Descent are
available as optimization methods.